Add use_kernels A/B toggle, perf logging, compile/compaction fixes, JAX efficiency mirror
- use_kernels master switch (MORPHConfig + base.yaml + _eager_flag): routes the
  4 fused attention kernels to their eager references and fused-CE to full-logits
  CE, for a same-architecture kernel-on/off A/B. x0-hoist + active-set stay on
  in both arms (bit-exact loop opts).
- train.py: log perf/tokens_per_sec, perf/peak_mem_alloc_mib, peak_reserved_mib.
- train.py: compile core MLPs with dynamic=True — active-set's variable sub-batch
  was triggering per-size recompiles; symbolic batch guard fixes it (warmup-only now).
- block_sparse.py topology_step: scatter src dtype = col_usage_count.dtype (robust
  if model cast to bf16; no-op in the fp32+autocast training path). Verified the
  prune->compact->sparse path + active-set variable-batch via ignore/verify_compaction.py.
- JAX/TPU mirror (morph/jax/): x0-hoist (bit-exact) + chunked CE (lax.scan,
  3.1GiB->96MiB; XLA does NOT auto-avoid the [N,V] materialisation) + lm_weight().
  Active-set intentionally NOT ported (XLA static shapes; existing where-mask is the
  TPU idiom). Verified on JAX-CPU (ignore/verify_jax_efficiency.py).

+++ morph/model/titans_core/block_sparse.py
@@ -1365,7 +1365,8 @@ class CMSBlockLinear(nn.Module):
             self.col_usage_count.scatter_add_(
                 0,
                 self.col_indices.flatten().long(),
-                torch.ones(self.R * self.K, device=self.col_indices.device, dtype=torch.float32)
+                torch.ones(self.R * self.K, device=self.col_indices.device,
+                           dtype=self.col_usage_count.dtype)  # match buffer dtype (robust if model cast)
             )
 
         # Invalidate column index cache since topology changed
